Refactor imports and remove unused network helpers for improved code clarity
