+++ apps/earn-protocol/app/earn/[network]/position/[vaultId]/[walletAddress]/page.tsx
@@ -1,6 +1,6 @@
 import { Text } from '@summerfi/app-earn-ui'
-import { humanNetworktoSDKNetwork, type SDKNetwork } from '@summerfi/app-types'
-import { parseServerResponseToClient } from '@summerfi/app-utils'
+import { type SDKNetwork } from '@summerfi/app-types'
+import { humanNetworktoSDKNetwork, parseServerResponseToClient } from '@summerfi/app-utils'
 import { type IArmadaPosition } from '@summerfi/sdk-client-react'
 
 import { getUserActivity } from '@/app/server-handlers/sdk/get-user-activity'
