+++ apps/earn-protocol/app/server-handlers/sdk/get-user-position.ts
@@ -1,9 +1,9 @@
 import { type SDKNetwork } from '@summerfi/app-types'
+import { subgraphNetworkToId } from '@summerfi/app-utils'
 import { type IArmadaPosition } from '@summerfi/sdk-client-react'
 import { Address, getChainInfoByChainId, User, Wallet } from '@summerfi/sdk-common'
 
 import { backendSDK } from '@/app/server-handlers/sdk/sdk-backend-client'
-import { subgraphNetworkToId } from '@/helpers/network-helpers'
 
 export async function getUserPosition({
   network,
