+++ apps/earn-protocol/hooks/use-client.ts
@@ -1,11 +1,9 @@
 import { useEffect, useMemo, useState } from 'react'
 import { useChain, useUser } from '@account-kit/react'
 import { type SDKVaultishType } from '@summerfi/app-types'
-import { ten } from '@summerfi/app-utils'
+import { subgraphNetworkToSDKId, ten } from '@summerfi/app-utils'
 import BigNumber from 'bignumber.js'
 import { createPublicClient, createWalletClient, custom, erc20Abi } from 'viem'
-
-import { subgraphNetworkToSDKId } from '@/helpers/network-helpers'
 
 export const useClient = ({ vault }: { vault?: SDKVaultishType }) => {
   const user = useUser()
