+++ packages/app-utils/src/index.ts
@@ -49,3 +49,11 @@ export {
 export { toggleArrayItem } from './helpers/toggle-array-item'
 export { simpleSort, SortDirection } from './helpers/simple-sort'
 export { getPastTimestamp } from './helpers/get-past-timestamp'
+export {
+  isSupportedSDKChain,
+  sdkNetworkToHumanNetwork,
+  humanNetworktoSDKNetwork,
+  networkNameToSDKNetwork,
+  subgraphNetworkToId,
+  subgraphNetworkToSDKId,
+} from './helpers/earn-network-tools'
